requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
filelock==3.18.0
flask==2.3.3
flask-cors==4.0.0
fsspec==2025.5.1
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.62.3
gunicorn==21.2.0
huggingface-hub==0.33.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.1
markdown==3.5.1
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numpy==2.3.0
packaging==25.0
pillow==11.2.1
promptfoo==0.1.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1-modules==0.4.2
python-dotenv==1.0.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.9.0
soupsieve==2.7
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typing-extensions==4.14.0
urllib3==2.4.0
werkzeug==2.3.7
